Bug #1160802:  Can't build Zope on Windows w/ 2.4.1c1.
MSVCCompiler.initialize():  set self.initialized to True, as suggested
by AMK.  Else we keep growing the PATH endlessly, with each new C
extension built, until putenv() complains.

This doesn't appear to be an issue on the HEAD (MSVCCompiler initializes
itself via __init__() on the HEAD).

Also added a "2.4.1c2" section to NEWS.  Not meant to imply that Anthony
will do a 2.4.1c2 release, just needed to a place to put the news about the
MSVCCompiler bugfix.