Fix bug #1000914 (again).
This patches a file that is generated by bgen, however the code is now the
same as a current copy of bgen would generate.  Without this patch most types
in the Carbon.CF module are unusable.

I haven't managed to coax bgen into generating a complete copy of _CFmodule.c
yet :-(, hence the manual patching.