#19662: Make requirement to support arbitrary keywords explicit.
When not using decode_data=True, smtpd may provide keyword arguments to the
process_message user-implemented method.  This doc update is intended to make
it clear that arbitrary keywords must be supported, so that we can add
additional features in the future by just adding keywords to the
process_message call.