Revert the change we made to _MAGIC_BYTES in hg:b9f656bc609c.
We turned this generator into a list comprehension way back in our first
commit because generators were crashing.  That turned out to be because we
were intercepting Py_CLEAR() and only calling it if we weren't a parallel
context.  We fixed that (much later!), so we can return this back to a
generator to reduce noise in the diffs against 3.3.5.