Remove our two Py_DECREF() calls in PyUnicode_InternInPlace() and replace with the original Py_REFCNT(s) -= 2.
I replaced the latter with the former very early on in PyParallel's
development, as part of the reference counting work.  The assumption was
"we shouldn't be manipulating reference counts directly, because parallel
threads need to avoid all reference count manipulation".  However, this is
a method that deals with unicode interning, and the 'Py_REFCNT(s) -= 2' had
a very specific purpose.  There isn't actually an issue regarding parallel
threads as this method has a check for Py_PXCTX() up the top and returns if
we're parallel.

This fixes some pretty obscure crashes I was running into.